fix Deadlock when activating the form sidebar
The form sidebar was created in parallel with some WollMux events which caused a deadlock on the TextDocumentModel.
So the form sidebar is no activated with an event on its own.
